Move all `bosh_common` code under `Bosh::Common`
