+++ src/bosh-director/lib/bosh/director/blobstores.rb
@@ -17,7 +17,7 @@ module Bosh::Director
       bare_client = create(provider_string, options)
 
       sha1_client = Blobstore::Sha1VerifyingClientWrapper.new(bare_client, Bosh::Director::Config.logger)
-      retry_config = Bosh::Retryable.new(tries: 6, sleep: 2.0, on: [Blobstore::BlobstoreError])
+      retry_config = Bosh::Common::Retryable.new(tries: 6, sleep: 2.0, on: [Blobstore::BlobstoreError])
       Blobstore::RetryableClientWrapper.new(sha1_client, retry_config)
     end
