+++ src/bosh-director/lib/bosh/director/thread_pool.rb
@@ -1,5 +1,5 @@
 module Bosh::Director
-  class ThreadPool < Bosh::ThreadPool
+  class ThreadPool < Bosh::Common::ThreadPool
     def initialize(**options)
       options[:logger] ||= Config.logger
       super(**options)
