+++ src/bosh-director/spec/unit/bosh/director/blobstore/retryable_client_wrapper_spec.rb
@@ -4,7 +4,7 @@ module Bosh::Director::Blobstore
   describe RetryableClientWrapper do
     subject { described_class.new(wrapped_client, retryable) }
     let(:wrapped_client) { instance_double('Bosh::Director::Blobstore::Client') }
-    let(:retryable)      { Bosh::Retryable.new(tries: 2, sleep: 0, on: [BlobstoreError]) }
+    let(:retryable)      { Bosh::Common::Retryable.new(tries: 2, sleep: 0, on: [BlobstoreError]) }
 
     it_calls_wrapped_client_methods(except: [:get])
