+++ src/bosh_common/lib/bosh/common/core_ext/kernel.rb
